+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new `defer()` helper.
 - Added `Cache::flexible()` method to add SWR support to the cache.
 - Added dynamic creation of post type/taxonomy factories.
+- Added `Reset_Server` trait to reset the server between tests.
+- Add `with_https()` to control if the request being tested is over HTTPS.
 
 ### Changed